- Integrated 64bit compatibility patch in utImage.cpp (thanks to mb108) - enabled extern "C" linking under Linux (thanks to Codepoet) - added EffectResParams to allow query of MaxLife and MinLife (thanks to swiftcoder)
git-svn-id: https://horde3d.svn.sourceforge.net/svnroot/horde3d/trunk@6 768fa460-af4a-0410-8817-c9c6fcdc2709

+++ Horde3D/Source/Horde3D Engine/utImage.cpp
@@ -1257,7 +1257,7 @@ static int process_frame_header(int scan)
          }
          return e("outofmem", "Out of memory");
       }
-      img_comp[i].data = (uint8*) (((int) img_comp[i].raw_data + 15) & ~15);
+	  img_comp[i].data = (uint8*) (((size_t) img_comp[i].raw_data + 15) & ~15);
       img_comp[i].linebuf = NULL;
    }
